+++ include/Grafo_matriz.h
@@ -27,9 +27,9 @@ class Grafo_matriz : public Grafo
         /**Destrutor da classe Grafo matriz */
         ~Grafo_matriz();
         /**Insere vértice com peso "val" */
-        void insereVertice(int val);
+        void insereVertice(float val);
         /**Insere aresta ao indicar o vértice de origem, o vértice de destino e o peso da aresta */
-        void insereAresta(int origem, int destino, int val);
+        void insereAresta(int origem, int destino, float val);
         /**Remove uma aresta ao passar o vértice de origem e qual o vértice que deve ser removido */
         void removeAresta(int i, int j);
